finance-tracker/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
plaid-python>=24.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
